+++ app/Controllers/CommentController.php
@@ -3,41 +3,93 @@ namespace App\Controllers;
 use App\Interfaces\CommentRepositoryInterface;
 use Classes\Comment;
 
+
+/**
+ * Class CommentController
+ *
+ * This controller manages actions related to comments, including listing, adding, 
+ * and deleting comments associated with news articles. It interacts with a 
+ * CommentRepository to perform database operations on the `comment` table.
+ */
 class CommentController
 {
 	private $commentRepository;
 	private static $instance = null;
 
-	private function __construct(CommentRepositoryInterface $commentRepositoryInterface)
+	public function __construct(CommentRepositoryInterface $commentRepositoryInterface)
 	{
 
 		$this->commentRepository = $commentRepositoryInterface;
 	}
-
+	/**
+	 * Create a new instance of the class if it doesn't already exist.
+	 * This implements the Singleton pattern, ensuring only one instance of the class.
+	 *
+	 * @param \App\Interfaces\CommentRepositoryInterface $commentRepositoryInterface
+	 *        An instance of the CommentRepositoryInterface to manage comment data.
+	 * @return object
+	 *         The singleton instance of this class.
+	 * TODO :: for foture, if application becomes big, you may use Dependency Inversion Pattern
+	 * as to it will be more advantagous for testing and flexibility 
+	 */
 	public static function getInstance(CommentRepositoryInterface $commentRepositoryInterface)
 	{
 		if (null === self::$instance) {
-			$c = __CLASS__;
+
 			self::$instance = new self($commentRepositoryInterface);
+
 		}
+
 		return self::$instance;
 	}
-
+	/**
+	 * list all comments
+	 * @return mixed
+	 */
 	public function listComments()
 	{
-		return $this->commentRepository->listAll();
-	}
+		try {
 
-	public function addCommentForNews($body, $newsId)
-	{
-		$comment = new Comment();
-		$comment->setBody($body)
-			->setNewsId($newsId);
-		return $this->commentRepository->add($comment);
-	}
+			return $this->commentRepository->listAll();
 
-	public function deleteComment($id)
+		} catch (\Throwable $th) {
+			// TODO :: Return a custom message you want here.
+		}
+
+	}
+	/**
+	 * Add a new comment record for an news article
+	 * @param string $body news article comment context
+	 * @param mixed $newsId news id
+	 * @return mixed
+	 */
+	public function addCommentForNews(string $body, int $newsId)
 	{
-		return $this->commentRepository->deleteById($id);
+		try {
+			$comment = new Comment($newsId, $body);
+
+			return $this->commentRepository->add($comment);
+
+		} catch (\Throwable $th) {
+			// TODO :: Return a custom message you want here.
+		}
+
+	}
+	/**
+	 * Delete a comment by its id
+	 * @param int $id
+	 * @return mixed
+	 */
+	public function deleteComment(int $id)
+	{
+		try {
+
+			return $this->commentRepository->deleteById($id);
+
+		} catch (\Throwable $th) {
+			// TODO :: Return a custom message you want here.
+			throw $th;
+		}
+
 	}
 }