+++ app/Services/ViewRendererService.php
@@ -5,6 +5,7 @@ class ViewRendererService
     public static function render($view, $data = [])
     {
         extract($data);
+
         include __DIR__ . '/../views/' . $view . '.php';
     }
 }